Move section to article package
