+++ backend/src/main/kotlin/com/jetbrains/life_science/article/section/dto/SectionDTO.kt
@@ -1,0 +1,15 @@
+package com.jetbrains.life_science.article.section.dto
+
+import javax.validation.constraints.NotBlank
+import javax.validation.constraints.Positive
+
+data class SectionDTO(
+
+    @field:NotBlank
+    val name: String,
+
+    @field:Positive
+    val articleVersionId: Long,
+
+    val description: String = ""
+)
