+++ backend/src/main/kotlin/com/jetbrains/life_science/article/section/search/repository/SectionSearchUnitRepository.kt
@@ -1,0 +1,6 @@
+package com.jetbrains.life_science.article.section.search.repository
+
+import com.jetbrains.life_science.article.section.search.SectionSearchUnit
+import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
+
+interface SectionSearchUnitRepository : ElasticsearchRepository<SectionSearchUnit, Long>
