+++ backend/src/main/kotlin/com/jetbrains/life_science/article/section/service/SectionCreationInfo.kt
@@ -1,0 +1,12 @@
+package com.jetbrains.life_science.article.section.service
+
+import com.jetbrains.life_science.article.version.entity.ArticleVersion
+
+interface SectionCreationInfo {
+
+    val article: ArticleVersion
+
+    val name: String
+
+    val description: String
+}
