+++ backend/src/main/kotlin/com/jetbrains/life_science/article/section/service/SectionInfo.kt
@@ -1,0 +1,12 @@
+package com.jetbrains.life_science.article.section.service
+
+interface SectionInfo {
+
+    val id: Long
+
+    val name: String
+
+    val description: String
+
+    val articleVersionId: Long
+}
