+++ backend/src/main/kotlin/com/jetbrains/life_science/article/section/service/SectionUpdateInfo.kt
@@ -1,0 +1,10 @@
+package com.jetbrains.life_science.article.section.service
+
+interface SectionUpdateInfo {
+
+    val id: Long
+
+    val name: String
+
+    val description: String
+}
