+++ backend/src/main/kotlin/com/jetbrains/life_science/article/section/view/SectionView.kt
@@ -1,0 +1,9 @@
+package com.jetbrains.life_science.article.section.view
+
+import com.jetbrains.life_science.content.view.ContentView
+
+class SectionView(
+    val name: String,
+    val description: String?,
+    val contents: List<ContentView>
+)
