+++ backend/src/main/kotlin/com/jetbrains/life_science/article/version/entity/ArticleVersion.kt
@@ -1,6 +1,6 @@
 package com.jetbrains.life_science.article.version.entity
 
-import com.jetbrains.life_science.section.entity.Section
+import com.jetbrains.life_science.article.section.entity.Section
 import com.jetbrains.life_science.article.master.entity.Article
 import com.jetbrains.life_science.user.entity.User
 import javax.persistence.*
