+++ backend/src/main/kotlin/com/jetbrains/life_science/article/version/service/ArticleVersionServiceImpl.kt
@@ -3,7 +3,7 @@ package com.jetbrains.life_science.article.version.service
 import com.jetbrains.life_science.article.master.service.ArticleService
 import com.jetbrains.life_science.exception.ArticleVersionNotFoundException
 import com.jetbrains.life_science.exception.PublishedVersionNotFoundException
-import com.jetbrains.life_science.section.service.SectionService
+import com.jetbrains.life_science.article.section.service.SectionService
 import com.jetbrains.life_science.article.version.adapter.SectionEmptyCreationToInfoAdapter
 import com.jetbrains.life_science.article.version.entity.ArticleVersion
 import com.jetbrains.life_science.article.version.entity.State
