+++ backend/src/main/kotlin/com/jetbrains/life_science/content/service/ContentService.kt
@@ -1,7 +1,7 @@
 package com.jetbrains.life_science.content.service
 
 import com.jetbrains.life_science.content.entity.Content
-import com.jetbrains.life_science.section.entity.Section
+import com.jetbrains.life_science.article.section.entity.Section
 
 interface ContentService {
