--- backend/src/main/kotlin/com/jetbrains/life_science/section/dto/SectionDTOToUpdateInfoAdapter.kt
@@ -1,17 +1,0 @@
-package com.jetbrains.life_science.section.dto
-
-import com.jetbrains.life_science.section.service.SectionUpdateInfo
-
-class SectionDTOToUpdateInfoAdapter(
-    val dto: SectionUpdateDTO
-) : SectionUpdateInfo {
-
-    override val id: Long
-        get() = dto.id
-
-    override val name: String
-        get() = dto.name
-
-    override val description: String
-        get() = dto.description
-}
