--- backend/src/main/kotlin/com/jetbrains/life_science/section/search/repository/SectionSearchUnitRepository.kt
@@ -1,6 +1,0 @@
-package com.jetbrains.life_science.section.search.repository
-
-import com.jetbrains.life_science.section.search.SectionSearchUnit
-import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
-
-interface SectionSearchUnitRepository : ElasticsearchRepository<SectionSearchUnit, Long>
