--- backend/src/main/kotlin/com/jetbrains/life_science/section/service/SectionInfo.kt
@@ -1,12 +1,0 @@
-package com.jetbrains.life_science.section.service
-
-interface SectionInfo {
-
-    val id: Long
-
-    val name: String
-
-    val description: String
-
-    val articleVersionId: Long
-}
