--- backend/src/main/kotlin/com/jetbrains/life_science/section/service/SectionUpdateInfo.kt
@@ -1,10 +1,0 @@
-package com.jetbrains.life_science.section.service
-
-interface SectionUpdateInfo {
-
-    val id: Long
-
-    val name: String
-
-    val description: String
-}
