--- backend/src/main/kotlin/com/jetbrains/life_science/section/view/SectionView.kt
@@ -1,9 +1,0 @@
-package com.jetbrains.life_science.section.view
-
-import com.jetbrains.life_science.content.view.ContentView
-
-class SectionView(
-    val name: String,
-    val description: String?,
-    val contents: List<ContentView>
-)
